english_phrasebook/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-rosetta==0.9.3
psycopg2-binary==2.8.5
Pillow==7.2.0
python-magic==0.4.18
eng-to-ipa==0.0.2
gunicorn==20.0.4